relocate json files to assets, makes more sense
